feat: refactor media components
Move carousel component to media/carousel directory:
- Moved carousel.tsx to media/carousel/carousel.tsx
- Created comprehensive documentation in README.md
- Added barrel exports via index.ts
- Updated backward compatibility exports with deprecation notices
- Created main media directory README.md and index.ts
- Updated COMPONENT-MOVEMENT-PLAN.md
