+++ src/components/ui/media/carousel/index.ts
@@ -1,0 +1,1 @@
+export * from "./carousel";
