+++ src/components/ui/media/index.ts
@@ -1,2 +1,3 @@
-export * from './avatar';
-export * from './image';
+export * from "./avatar";
+export * from "./image";
+export * from "./carousel";
